+++ src/Framework/FWFemMesh.cpp
@@ -306,7 +306,7 @@ bool FWFemMesh::CreatePHFromGR(){
 	std::vector<Vec3d> vtxsIn;
 	for(unsigned i=0; i<grMesh->vertices.size(); ++i) vtxsIn.push_back(grMesh->vertices[i]);
 	// swithes q+(半径/最短辺) (e.g. = q1.0~2.0) a 最大の体積 
-	sprTetgen(nVtxsOut, vtxsOut, nTetsOut, tetsOut, grMesh->vertices.size(), &vtxsIn[0], grMesh->faces.size(), &grMesh->faces[0], "pq2.1a0.002");//a0.3 //a0.003 
+	sprTetgen(nVtxsOut, vtxsOut, nTetsOut, tetsOut, (int)grMesh->vertices.size(), &vtxsIn[0], (int)grMesh->faces.size(), &grMesh->faces[0], "pq2.1a0.002");//a0.3 //a0.003 
 	
 	//	PHMesh用のディスクリプタpmdに値を入れていく
 	for(int i=0; i < nVtxsOut; i++){
@@ -533,7 +533,7 @@ void FWFemMesh::CreateGRFromPH(){
 					gmd.vertices.push_back(gmd.vertices[pv]);
 					gmd.texCoords.push_back(texCoord);
 					if (gmd.normals.size()) gmd.normals.push_back(normal);
-					gmd.faces[pf].indices[i] = gmd.vertices.size()-1;
+					gmd.faces[pf].indices[i] = (int)gmd.vertices.size()-1;
 					vertexIdMap.push_back(phMesh->surfaceVertices[pv]);
 				}
 			}else{	//	そうでなければ、直接代入
